+++ src/datasets/datasets.module.ts
@@ -13,6 +13,7 @@ import { PoliciesService } from "src/policies/policies.service";
 import { PoliciesModule } from "src/policies/policies.module";
 import { ElasticSearchModule } from "src/elastic-search/elastic-search.module";
 import { DatasetsV4Controller } from "./datasets.v4.controller";
+import { DatasetsAccessService } from "./datasets-access.service";
 
 @Module({
   imports: [
@@ -64,6 +65,6 @@ import { DatasetsV4Controller } from "./datasets.v4.controller";
   ],
   exports: [DatasetsService],
   controllers: [DatasetsController, DatasetsV4Controller],
-  providers: [DatasetsService, CaslAbilityFactory],
+  providers: [DatasetsService, CaslAbilityFactory, DatasetsAccessService],
 })
 export class DatasetsModule {}
