+++ src/datasets/datasets.v4.controller.ts
@@ -76,7 +76,6 @@ import { IncludeValidationPipe } from "./pipes/include-validation.pipe";
 import { PidValidationPipe } from "./pipes/pid-validation.pipe";
 import { FilterValidationPipe } from "./pipes/filter-validation.pipe";
 import { getSwaggerDatasetFilterContent } from "./types/dataset-filter-content";
-import { logger } from "@user-office-software/duo-logger";
 
 export interface IDatasetFiltersV4<T, Y = null> {
   where?: FilterQuery<T>;
@@ -301,13 +300,9 @@ export class DatasetsV4Controller {
           "A dataset with this this unique key already exists!",
         );
       } else {
-        logger.logException(
-          "Something went wrong while creating the dataset",
-          error,
-        );
-
         throw new InternalServerErrorException(
           "Something went wrong. Please try again later.",
+          { cause: error },
         );
       }
     }
